pyproject.toml
[project]
name = "eluent"
version = "0.0.2"
authors = [
  { name="Eachan Johnson", email="eachan.johnson@crick.ac.uk" },
]
description = " Chemistry-aware splitting of large datasets."
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }
keywords = [
  "ai",
  "data",
  "deep-learning",
  "machine-learning",
  "chemistry",
]

classifiers = [  
  
  "Development Status :: 3 - Alpha",

  # Indicate who your project is intended for
  "Intended Audience :: Science/Research",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",

  "License :: OSI Approved :: MIT License",

  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3 :: Only",
]

dependencies = [
  "carabiner-tools[pd,mpl]>=0.0.5.post3",
  "datasets>=3.0.0",
  "huggingface_hub",
  "numpy",
  "platformdirs>=4.4.0",
  "scipy",
  "tensorboard",
  "tqdm",
  "umap-learn", 
  "scipy", 
  "schemist>=0.0.6", 
  "tdigest>=0.5", 
  "faiss-cpu>=1.11",
]

[project.optional-dependencies]
splits_gpu = ["faiss-gpu"]
dev = ["pytest", "flake8"]

[project.urls]
"Homepage" = "https://github.com/scbirlab/eluent"
"Bug Tracker" = "https://github.com/scbirlab/duvidnn/eluent"

[project.scripts] 
eluent = "eluent.cli_module.cli:main"

[tool.setuptools]
#packages = ["eluent"]
# If there are data files included in your packages that need to be
# installed, specify them here.
package-data = {eluent = ["data/*.json"]}

[build-system]
# These are the assumed default build requirements from pip:
# https://pip.pypa.io/en/stable/reference/pip/#pep-517-and-518-support
requires = ["setuptools>=43.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[test]
deps = ["pytest>=7.4", "pytest-cov", "eluent[dev]"] 
